+++ src/frontend/apps/impress/src/features/left-panel/components/LeftPanelHeader.tsx
@@ -27,7 +27,7 @@ export const LeftPanelHeader = ({ children }: PropsWithChildren) => {
   useCmdK(openSearchModal);
   const { togglePanel } = useLeftPanelStore();
 
-  const { mutate: createDoc } = useCreateDoc({
+  const { mutate: createDoc, isPending: isCreatingDoc } = useCreateDoc({
     onSuccess: (doc) => {
       router.push(`/docs/${doc.id}`);
       togglePanel();
@@ -75,7 +75,9 @@ export const LeftPanelHeader = ({ children }: PropsWithChildren) => {
               )}
             </Box>
             {authenticated && (
-              <Button onClick={createNewDoc}>{t('New doc')}</Button>
+              <Button onClick={createNewDoc} disabled={isCreatingDoc}>
+                {t('New doc')}
+              </Button>
             )}
           </Box>
         </SeparatedSection>